Add sampling profiler implementation
As a tracing profiler decorator.
